UI changes, added WSTagField, added ViewAnimator, added IQKeyboardManager, added ProjectEditor, added DatabaseHandler
UI - added dark theme only - it's an app for devs and all devs love dark theme so there is on reason to have multiple themes.
WSTagField - pod for creating tags on textfield
ViewAnimator - pod for animations - to be used
IQKeyboardManager - useful keyboard manager that handles view raising above keyboard
ProjectEditor - screen for creating projects
DatabaseHandler - my database handler that I used in previous projects. It is designed to be used with firebase. This time, all supportive enums and extensions are in DBHandler for quick access. DatabaseDocument model is added that contains id of the document and any object in the document. This is to make referencing id's easier and to avoid having separate id in the model like Project.